Backport PR #17600: FIX: work with PyQt 5.15
Merge pull request #17600 from tacaswell/mnt_more_qt515_fixes

FIX: work with PyQt 5.15
Conflicts:
	lib/matplotlib/backends/backend_qt5.py
          - on this branch the blitting code is still in
            backend_qt5agg.py

+++ lib/matplotlib/backends/backend_qt5.py
@@ -452,7 +452,8 @@ class FigureCanvasQT(QtWidgets.QWidget, FigureCanvasBase):
             raise RuntimeError("Event loop already running")
         self._event_loop = event_loop = QtCore.QEventLoop()
         if timeout > 0:
-            timer = QtCore.QTimer.singleShot(timeout * 1000, event_loop.quit)
+            timer = QtCore.QTimer.singleShot(int(timeout * 1000),
+                                             event_loop.quit)
         event_loop.exec_()
 
     def stop_event_loop(self, event=None):
